feat/ Add logic to helper method to calculate FUV
A helper method calculates all entries for the FUV vector and returns
it to `decide()`. The method uses the fact that the default value of
boolean vector elements are `false`, so if false is ever generated from
the `PUM` we stop evaluating the remaining signals and return the FUV
since the `decide()` should print "NO" regardless of the remaining
elements, so they can be false.

We do not have to worry about potential null values as arrays of
primitive types can not be null and all arrays we have built hold the
primitive type `boolean` and not the object `Boolean`.